requirements-dev.txt
-r requirements.txt
pytest==8.1.1
pytest-django==4.8.0
mixer==7.2.2
coverage==7.4.4
django-debug-toolbar==4.3.0
